+++ R/board_github.R
@@ -17,7 +17,14 @@ github_headers <- function(board) {
   httr::add_headers(Authorization = paste("token", github_auth(board)))
 }
 
-board_initialize.github <- function(board, token = NULL, repo = NULL, path = "", branch = "master", overwrite = FALSE, ...) {
+board_initialize.github <- function(board,
+                                    token = NULL,
+                                    repo = NULL,
+                                    path = "",
+                                    branch = "master",
+                                    overwrite = FALSE,
+                                    host = "https://api.github.com",
+                                    ...) {
   if (!github_authenticated(board)) {
     if (is.null(token)) {
       stop("GitHub Personal Access Token must be specified with the 'token' parameter or with the 'GITHUB_PAT' ",
@@ -33,6 +40,7 @@ board_initialize.github <- function(board, token = NULL, repo = NULL, path = "",
   board$repo <- repo
   board$path <- if (!is.null(path) && nchar(path) > 0) paste0(path, "/") else ""
   board$branch <- branch
+  board$host <- host
 
   # check repo exists
   check_exists <- httr::GET(github_url(board, branch = NULL), github_headers(board))
@@ -53,7 +61,7 @@ board_initialize.github <- function(board, token = NULL, repo = NULL, path = "",
   board
 }
 
-github_update_index <- function(board, path, commit, operation, name = NULL, metadata = NULL, branch = board$branch) {
+github_update_temp_index <- function(board, path, commit, operation, name = NULL, metadata = NULL, branch = board$branch) {
   index_url <- github_url(board, branch = branch, "/contents/", board$path, "data.txt")
   response <- httr::GET(index_url, github_headers(board))
 
@@ -101,6 +109,17 @@ github_update_index <- function(board, path, commit, operation, name = NULL, met
   index_file <- tempfile(fileext = "yml")
   board_manifest_create(index, index_file)
 
+  list(
+    index_file = index_file,
+    sha = sha
+  )
+}
+
+github_update_index <- function(board, path, commit, operation, name = NULL, metadata = NULL, branch = board$branch) {
+  index <- github_update_temp_index(board, path, commit, operation, name = NULL, metadata = NULL, branch = board$branch)
+  index_file <- index$index_file
+  sha <- index$sha
+
   file_url <- github_url(board, branch = branch, "/contents/", board$path, "data.txt")
 
   base64 <- base64enc::base64encode(index_file)
@@ -118,6 +137,17 @@ github_update_index <- function(board, path, commit, operation, name = NULL, met
   }
 }
 
+github_delete_release <- function(board, release_tag) {
+  response <- httr::GET(github_url(board, branch = NULL, paste0("/releases/tags/", release_tag)), github_headers(board))
+  if (!httr::http_error(response)) {
+    release_id <- httr::content(response)$id
+    response <- httr::DELETE(github_url(board, branch = NULL, paste0("/releases/", release_id)), github_headers(board))
+    if (httr::http_error(response)) {
+      pin_log("Failed to delete release ", release_id)
+    }
+  }
+}
+
 github_create_release <- function(board, name) {
   index_url <- github_url(board, branch = NULL, "/commits/", board$branch)
   response <- httr::GET(index_url, github_headers(board))
@@ -126,10 +156,20 @@ github_create_release <- function(board, name) {
 
   release_url <- github_url(board, branch = NULL, "/releases")
 
+  release_tag <- release_name <- name
+  if (board_versions_enabled(board, TRUE)) {
+    release_tag <-  paste(name, version, sep = "-")
+    release_name <- paste(name, version)
+  }
+  else {
+    # attempt to delete existing release
+    github_delete_release(board, release_tag)
+  }
+
   release <- list(
-    tag_name = paste(name, version, sep = "-"),
+    tag_name = release_tag,
     target_commitish = board$branch,
-    name = paste(name, version),
+    name = release_name,
     body = paste0("Storage for resource '", name, "' which is too large to be stored as a GitHub file.")
   )
 
@@ -172,13 +212,137 @@ github_upload_content <- function(board, name, file, file_path, commit, sha, bra
                           branch = branch
                         ),
                         github_headers(board), encode = "json",
-                        http_utils_progress("up"))
+                        http_utils_progress("up", size = file.info(normalizePath(file_path))$size))
   upload <- httr::content(response)
 
   if (httr::http_error(response)) {
     pin_log("Failed to upload ", file, " response: ", upload)
     stop("Failed to upload ", file, " to ", board$repo, ": ", upload$message)
   }
+}
+
+github_upload_blob <- function(board, file, file_path, commit) {
+  blob_url <- github_url(board, branch = NULL, "/git/blobs")
+  pin_log("uploading ", file)
+
+  base64 <- base64enc::base64encode(file_path)
+  response <- httr::POST(blob_url,
+                         body = list(
+                           content = base64,
+                           encoding = "base64"
+                         ),
+                         github_headers(board), encode = "json",
+                         http_utils_progress("up", size = file.info(normalizePath(file_path))$size))
+  upload <- httr::content(response)
+
+  if (httr::http_error(response)) {
+    pin_log("Failed to upload ", file, " response: ", upload)
+    stop("Failed to upload ", file, " to ", board$repo, ": ", upload$message)
+  }
+
+  upload
+}
+
+github_create_tree <- function(board, tree_files, base_sha) {
+  tree_url <- github_url(board, branch = NULL, "/git/trees")
+  pin_log("creating tree")
+
+  response <- httr::POST(tree_url,
+                         body = list(
+                           base_tree = jsonlite::unbox(base_sha),
+                           tree = tree_files
+                         ),
+                         github_headers(board), encode = "json",
+                         http_utils_progress("up"))
+  upload <- httr::content(response)
+
+  if (httr::http_error(response)) {
+    pin_log("Failed to create tree, response: ", upload)
+    stop("Failed to create tree in ", board$repo, ": ", upload$message)
+  }
+
+  upload
+}
+
+github_create_commit <- function(board, tree_sha, base_sha, commit) {
+  commit_url <- github_url(board, branch = NULL, "/git/commits")
+  pin_log("creating commit")
+
+  response <- httr::POST(commit_url,
+                         body = list(
+                           message = jsonlite::unbox(commit),
+                           tree = jsonlite::unbox(tree_sha),
+                           parents = list(
+                             jsonlite::unbox(base_sha)
+                           )
+                         ),
+                         github_headers(board), encode = "json",
+                         http_utils_progress("up"))
+  upload <- httr::content(response)
+
+  if (httr::http_error(response)) {
+    pin_log("Failed to create commit, response: ", upload)
+    stop("Failed to create commit in ", board$repo, ": ", upload$message)
+  }
+
+  upload
+}
+
+github_update_head <- function(board, branch, commit_sha) {
+  ref_url <- github_url(board, branch = NULL, paste0("/git/refs/heads/", branch))
+  pin_log("updating head")
+
+  response <- httr::VERB("PATCH",
+                         ref_url,
+                         body = list(
+                           sha = jsonlite::unbox(commit_sha),
+                           force = jsonlite::unbox(FALSE)
+                         ),
+                         github_headers(board), encode = "json",
+                         http_utils_progress("up"))
+  upload <- httr::content(response)
+
+  if (httr::http_error(response)) {
+    pin_log("Failed to update branch, reponse: ", upload)
+    stop("Failed to update branch ", branch, ": ", upload$message)
+  }
+
+  upload
+}
+
+github_refs_head <- function(board, branch) {
+  ref_url <- github_url(board, branch = NULL, paste0("/git/ref/heads/", branch))
+  response <- httr::GET(ref_url, github_headers(board))
+
+  reference <- httr::content(response)
+
+  if (httr::http_error(response)) {
+    pin_log("Failed to retrieve branch, reponse: ", response)
+    stop("Failed to retrieve branch ", branch, ": ", reference$message)
+  }
+
+  reference
+}
+
+github_files_commit <- function(board, upload_files, branch, commit) {
+  tree_files <- list()
+  for (file in names(upload_files)) {
+
+    result <- github_upload_blob(board, file, upload_files[[file]], commit)
+
+    tree_files[[file]] <- list(
+      path = jsonlite::unbox(file),
+      mode = jsonlite::unbox('100644'),
+      type = jsonlite::unbox('blob'),
+      sha = jsonlite::unbox(result$sha)
+    )
+  }
+
+  head_sha <- github_refs_head(board, branch)
+  tree_result <- github_create_tree(board, unname(tree_files), head_sha$object$sha)
+  commit_result <- github_create_commit(board, tree_result$sha, head_sha$object$sha, commit)
+
+  github_update_head(board, branch, commit_result$sha)
 }
 
 board_pin_create.github <- function(board, path, name, metadata, ...) {
@@ -206,50 +370,52 @@ board_pin_create.github <- function(board, path, name, metadata, ...) {
     file.copy(path, bundle_path)
   }
 
-  dir_shas <- NULL
-  dir_url <- github_url(board, branch = branch, "/contents/", board$path, name)
-  dir_response <- httr::GET(dir_url, github_headers(board))
-  if (!httr::http_error(dir_response)) {
-    dir_shas <- httr::content(dir_response)
-  }
-
   release <- NULL
   release_map <- list()
   upload_files <- dir(bundle_path, recursive = TRUE)
 
-  # data.txt must be last to fix release paths
-  upload_files <- c(Filter(function(e) e != "data.txt", upload_files), ifelse("data.txt" %in% upload_files, "data.txt", NULL))
-
+  # first upload large files
   for (file in upload_files) {
-    commit <- if (is.null(list(...)$commit)) paste("update", name) else list(...)$commit
-    named_sha <- Filter(function(e) identical(e$path, paste0(board$path, file.path(name, file))), dir_shas)
-    sha <- if (length(named_sha) > 0) named_sha[[1]]$sha else NULL
-
     file_path <- file.path(bundle_path, file)
-
-    if (identical(file, "data.txt")) {
-      datatxt <- yaml::read_yaml(file_path, eval.expr = FALSE)
-      datatxt$path <- sapply(datatxt$path, function(e) { if(e %in% names(release_map)) release_map[[e]] else e })
-      yaml::write_yaml(datatxt, file_path)
-    }
 
     if ((file.info(file_path)$size > getOption("pins.github.release", 25) * 10^6 || release_storage) && !identical(file, "data.txt")) {
       if (is.null(release)) release <- github_create_release(board, name)
       download_url <- github_upload_release(board, release, name, file, file_path)
       release_map[[file]] <- download_url
     }
-    else {
-      github_upload_content(board, name, file, file_path, commit, sha, branch)
-    }
   }
 
+  # remove uploaded files
+  upload_files <- Filter(function(e) !e %in% names(release_map), upload_files)
+
+  # update data.txt with large files
+  if ("data.txt" %in% upload_files) {
+    file_path <- file.path(bundle_path, "data.txt")
+
+    datatxt <- suppressWarnings(yaml::read_yaml(file_path, eval.expr = FALSE))
+
+    datatxt$path <- sapply(datatxt$path, function(e) { if(e %in% names(release_map)) release_map[[e]] else e })
+    yaml::write_yaml(datatxt, file_path)
+  }
+
+  # create upload definition of remote-path/local-path
+  upload_defs <- file.path(bundle_path, upload_files)
+  names(upload_defs) <- paste0(board$path, name, "/", upload_files)
+
+  # update local index
   if (update_index) {
     index_path <- paste0(board$path, name)
 
-    github_update_index(board, index_path, commit, operation = "create",
-                        name = name, metadata = metadata, branch = branch)
+    index <- github_update_temp_index(board, index_path, commit, operation = "create",
+                                            name = name, metadata = metadata, branch = branch)
+
+    remote_index <- paste0(board$path, "data.txt")
+    upload_defs[[remote_index]] <- index$index_file
   }
 
+  # add remaining files in a single commit
+  commit <- if (is.null(list(...)$commit)) paste("update", name) else list(...)$commit
+  github_files_commit(board, upload_defs, branch, commit)
 }
 
 board_pin_find.github <- function(board, text, ...) {
@@ -303,7 +469,7 @@ board_pin_find.github <- function(board, text, ...) {
 github_url <- function(board, branch = board$branch, ...) {
   args <- list(...)
 
-  url <- paste0("https://api.github.com/repos/", board$repo, paste0(args, collapse = ""))
+  url <- paste0(board$host, "/repos/", board$repo, paste0(args, collapse = ""))
   if (!is.null(branch))
     url <- paste0(url, "?ref=", branch)
 
@@ -354,7 +520,7 @@ github_branches_create <- function(board, new_branch, base_branch) {
 github_content_url <- function(board, branch = board$branch, ...) {
   args <- list(...)
 
-  url <- paste0("https://api.github.com/repos/", board$repo, "/contents/", paste0(board$path, paste0(args, collapse = "")))
+  url <- paste0(board$host, "/repos/", board$repo, "/contents/", paste0(board$path, paste0(args, collapse = "")))
   if (!is.null(branch))
     url <- paste0(url, "?ref=", branch)
 
@@ -381,11 +547,17 @@ github_download_files <- function(index, temp_path, board) {
   }
 }
 
-board_pin_get.github <- function(board, name, extract = NULL, ...) {
+board_pin_get.github <- function(board, name, extract = NULL, version = NULL, ...) {
   branch <- if (is.null(list(...)$branch)) board$branch else list(...)$branch
+  subpath <- name
+
+  if (!is.null(version)) {
+    branch <- version
+    subpath <- file.path(name, pin_versions_path_name(), version)
+  }
 
   base_url <- github_raw_url(board, branch = branch, board$path, name, "/data.txt")
-  local_path <- pin_download(base_url, name, board$name, headers = github_headers(board))
+  local_path <- pin_download(base_url, name, board$name, headers = github_headers(board), subpath = subpath)
 
   if (file.exists(file.path(local_path, "data.txt"))) {
     index_path <- pin_manifest_download(local_path)
@@ -401,7 +573,7 @@ board_pin_get.github <- function(board, name, extract = NULL, ...) {
       else {
         file_url <- github_raw_url(board, branch = branch, board$path, name, "/", file)
       }
-      pin_download(file_url, name, board$name, headers = headers, extract = identical(extract, TRUE))
+      pin_download(file_url, name, board$name, headers = headers, extract = identical(extract, TRUE), subpath = subpath)
     }
 
     local_path
@@ -458,9 +630,36 @@ board_pin_remove.github <- function(board, name, ...) {
       stop("Failed to delete ", name, " from ", board$repo, ": ", deletion$message)
   }
 
+  if (!board_versions_enabled(board, TRUE)) {
+    github_delete_release(board, name)
+  }
+
   if (update_index) github_update_index(board, paste0(board$path, name), commit, operation = "remove")
 }
 
 board_browse.github <- function(board) {
   utils::browseURL(paste0("https://github.com/", board$repo, "/tree/",board$branch, "/", board$path))
+}
+
+board_pin_versions.github <- function(board, name, ...) {
+  branch <- if (is.null(list(...)$branch)) board$branch else list(...)$branch
+
+  path <- paste0(board$path, name)
+  response <- httr::GET(github_url(board, "/commits", branch = NULL,
+                                   sha = "?per_page=100&sha=", branch,
+                                   "&path=", path),
+                      github_headers(board))
+
+  commits <- httr::content(response)
+
+  if (httr::http_error(response))
+    stop("Failed to retrieve commits from ", board$repo, ": ", commits$message)
+
+  data.frame(
+    version = sapply(commits, function(e) e$sha),
+    created = sapply(commits, function(e) e$commit$author$date),
+    author = sapply(commits, function(e) e$author$login),
+    message = sapply(commits, function(e) e$commit$message),
+    stringsAsFactors = FALSE) %>%
+    format_tibble()
 }